ascii-art-python/pyproject.toml
[project]
name = "ascii-art-generator"
version = "1.0.0"
description = "将图片转换为ASCII艺术的生成器"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "mcp[cli]>=1.2.0",
    "supabase>=2.4.0",
]

[tool.setuptools]
py-modules = ["main", "ascii_art_server", "test_mcp_server"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

